Rename currencies file to getCurrency
